+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2022-04-20 :sparkles:
+- Adds features and a CLI to generate artifacts from OpenAPI Documentation
+  files (markdown for MkDocs and PyMdown extensions, PlantUML class diagrams
+  from components schemas)
+
 ## [0.1.6] - 2021-11-17 :gem:
 - Adds `py.typed` file
 - Add `Python 3.10` to the GitHub Workflow
@@ -27,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Corrects a bug forcing `camelCase` on examples objects handled as dataclasses
 - Adds base64 ValueFormat to the v3 enum
+
 ## [0.1.2] - 2021-05-03 :notes:
 
 - Adds a changelog